# src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Enterprise Flows Repository - Infrastructure",
        "version": "1.0.0",
        "description": "API for enterprise architecture.\nInfrastructure description.",
        "contact": {
            "name": "Emmanuel LESNE",
            "url": "https://github.com/MiddlewareSolutions/EnterpriseFlowRepository/",
            "email": "emmanuel.lesne@middleware-solutions.fr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        }
    },
    "servers": [
        {
            "url": "http://www.middleware-solutions.com/{stage}/api/ea/infrastructure/1",
            "variables": {
                "stage": {
                    "default": "dev",
                    "description": "Snapshot development"
                }
            }
        }
    ],
    "paths": {
...
        "/server": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Server"
                                    }
                                }
                            }
                        },
                        "description": "List of servers."
                    }
                },
                "operationId": "listServers",
                "description": "list all servers"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Server"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Modification accepted."
                    }
                },
                "operationId": "addServer",
                "description": "add a server"
            }
        },
        "/server/{id}": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Server"
                                }
                            }
                        },
                        "description": "Description of a server."
                    }
                },
                "operationId": "getOneServer",
                "description": "Get description of a server."
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Server"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Modification accepted."
                    }
                },
                "operationId": "modifyOneServer",
                "description": "Modify this server."
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "Deletion accepted."
                    }
                },
                "operationId": "deleteOneServer",
                "description": "Delete this server"
            },
            "parameters": [
                {
                    "name": "id",
                    "description": "ID of element",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                },
                {
                    "name": "version",
                    "description": "version to retreive",
                    "schema": {
                        "type": "string"
                    },
                    "in": "query"
                }
            ]
        },
        "/endpoint": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EndPoint"
                                    }
                                }
                            }
                        },
                        "description": "List endpoints."
                    }
                },
                "operationId": "listEndPoints",
                "description": "list all endpoints"
            },
            "post": {
                "requestBody": {
                    "description": "New endpoint.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EndPoint"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Endpoint added."
                    }
                },
                "operationId": "addEndPoint",
                "description": "add a endpoint"
            }
        },
        "/endpoint/{id}": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EndPoint"
                                }
                            }
                        },
                        "description": "Endpoint found."
                    }
                },
                "operationId": "getOneEndPoint",
                "description": "Get description of a endpoint."
            },
            "post": {
                "requestBody": {
                    "description": "Modifications for this endpoint.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EndPoint"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Modification accepted."
                    }
                },
                "operationId": "modifyOneEndPoint",
                "description": "Modify this endpoint."
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "Deletion accepted."
                    }
                },
                "operationId": "deleteOneEndPoint",
                "description": "Delete this endpoint"
            },
            "parameters": [
                {
                    "name": "id",
                    "description": "ID of element",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                },
                {
                    "name": "version",
                    "description": "version to retreive.",
                    "schema": {
                        "type": "string"
                    },
                    "in": "query"
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
            "Element": {
                "title": "Root Type for Element",
                "description": "The root of the Element type's schema.",
                "type": "object",
                "properties": {
                    "id": {
                        "description": "ID of this element.",
                        "type": "string"
                    },
                    "class": {
                        "description": "Type of element",
                        "type": "string"
                    },
                    "link": {
                        "description": "Link to this element",
                        "type": "string"
                    }
                },
                "example": {
                    "id": "ABC",
                    "class": "DataCenter",
                    "link": "/datacenter/ABC"
                }
            },
...
            "Server": {
                "description": "description of",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Base"
                    },
                    {
                        "type": "object"
                    }
                ],
                "properties": {
                    "network": {
                        "$ref": "#/components/schemas/Element",
                        "description": "run in this network."
                    },
                    "CPU": {
                        "format": "double",
                        "description": "Ressources available.",
                        "type": "number"
                    },
                    "RAM": {
                        "format": "double",
                        "description": "Ressources available.",
                        "type": "number"
                    },
                    "softwares": {
                        "description": "List of softwares installed.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Element"
                        }
                    },
                    "endpoints": {
                        "description": "endpoints of this server.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EndPoint"
                        }
                    },
                    "serversVirtualized": {
                        "description": "Servers virtualized",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Element"
                        }
                    }
                },
                "example": {
                    "network": {
                        "id": "ABC",
                        "class": "Network",
                        "link": "/datacenter/123/zone/12/network/ABC"
                    },
                    "CPU": 83,
                    "RAM": 88.34,
                    "softwares": [
                        {
                            "id": "1234",
                            "class": "Software",
                            "link": "/software/1234"
                        },
                        {
                            "id": "1235",
                            "class": "Software",
                            "link": "/software/1235"
                        }
                    ],
                    "endpoints": [
                        {
                            "IP": "10.0.0.1",
                            "URL": "http://monsite.com",
                            "port": 8080
                        },
                        {
                            "IP": "10.0.0.2",
                            "URL": "http://monsite2.com",
                            "port": 8081
                        }
                    ],
                    "serversVirtualized": [
                        {
                            "id": "1732",
                            "class": "Server",
                            "link": "/server/1732"
                        }
                    ]
                }
            },
...
            "EndPoint": {
                "description": "description of",
                "required": [
                    "IP"
                ],
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Base"
                    },
                    {
                        "type": "object"
                    }
                ],
                "properties": {
                    "IP": {
                        "description": "IP of this endpoint.",
                        "type": "string"
                    },
                    "URL": {
                        "description": "URL of this endpoint.",
                        "type": "string"
                    },
                    "port": {
                        "description": "port of this endpoint.",
                        "type": "integer"
                    }
                }
            },
...
            "Base": {
                "description": "description of",
                "required": [
                    "temporal",
                    "deleted"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Base"
                    },
                    {
                        "type": "object"
                    }
                ],
                "properties": {
                    "id": {
                        "description": "ID of element.",
                        "type": "string"
                    },
                    "name": {
                        "description": "name of element.",
                        "type": "string"
                    },
                    "description": {
                        "description": "Textual description.",
                        "type": "string"
                    },
                    "temporal": {
                        "$ref": "#/components/schemas/Temporal",
                        "description": "Temporal informations of element."
                    },
                    "deleted": {
                        "description": "Is it deleted ?",
                        "type": "boolean"
                    },
                    "tags": {
                        "description": "Tags on this element.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                },
                "example": {
                    "id": "123",
                    "name": "Server 1",
                    "description": "linux server N°1",
                    "temporal": {
                        "start": "some text",
                        "end": "some text",
                        "obsolete": "some text",
                        "version": "some text",
                        "previous": {
                            "id": "some text",
                            "class": "some text",
                            "link": "some text"
                        },
                        "element": {
                            "id": "some text",
                            "class": "some text",
                            "link": "some text"
                        },
                        "uuid": "some text"
                    },
                    "deleted": true,
                    "tags": [
                        {
                            "name": "some text",
                            "value": "some text"
                        },
                        {
                            "name": "some text",
                            "value": "some text"
                        }
                    ]
                }
            },
            "Temporal": {
                "title": "Root Type for Validity",
                "description": "The root of the Validity type's schema.",
                "required": [
                    "element",
                    "uuid",
                    "start"
                ],
                "type": "object",
                "properties": {
                    "start": {
                        "format": "date-time",
                        "description": "start of life",
                        "type": "string"
                    },
                    "end": {
                        "format": "date-time",
                        "description": "End of life",
                        "type": "string"
                    },
                    "obsolete": {
                        "format": "date-time",
                        "description": "start of obsolecence.",
                        "type": "string"
                    },
                    "version": {
                        "description": "Version of element.",
                        "type": "string"
                    },
                    "previous": {
                        "$ref": "#/components/schemas/Element",
                        "description": "Previsous temporal"
                    },
                    "element": {
                        "$ref": "#/components/schemas/Element",
                        "description": "Element of this temporal description."
                    },
                    "uuid": {
                        "description": "Unique id",
                        "type": "string"
                    }
                },
                "example": {
                    "start": "2018-02-10T09:30Z",
                    "end": "2018-02-10T09:30Z",
                    "obsolete": "2018-02-10T09:30Z",
                    "version": "1.0.1",
                    "previous": {
                        "id": "123",
                        "class": "Server",
                        "link": "/server/123?version=1.0.0"
                    },
                    "element": {
                        "id": "124",
                        "class": "Server",
                        "link": "/server/123"
                    },
                    "uuid": "abc-123-def-456"
                }
            },
...
            "Tag": {
                "description": "Tag to filter and specify elements.",
                "required": [
                    "value",
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "name of tag",
                        "type": "string"
                    },
                    "value": {
                        "description": "value of tag",
                        "type": "string"
                    }
                },
                "example": {
                    "name": "project",
                    "value": "ABC"
                }
            },
...
        },
...
    },
...
}